Fix :: 500 에러 수정

+++ src/main/kotlin/com/example/wunyo/domain/hobby/service/QueryHobbyRankingService.kt
@@ -11,7 +11,7 @@ class QueryHobbyRankingService(
 ) {
 
     fun getHobbyRankingList(): QueryHobbyRankingList {
-        val hobbyAllCount = hobbyRepository.findAll().count()
+        val hobbyAllCount = hobbyRepository.findAll().count().toDouble()
         val hobbyCountMap = hobbyRepository.countHobby() // key : 취미 이름, value : 취미 수
 
         val response = hobbyCountMap.map {
